requirements.txt
groundx-python-sdk==1.3.2
openai==0.28.1
python-telegram-bot==20.6
